google-chrome-stable does not exist, replace with chromium
I saw a reference to Google Chrome stable being installed in the dockerfile, but ultimately the only binary that’s showed up in my fresh installs in /usr/bin/chromium.

I was getting 500s when attempting to generate PDFs. After changing this and restarting, I now see PDF previews.

If this PR gets treated as an issue and there’s a way better/different fix, that’s cool too.

Signed-off-by: Aki <hi@akiro.se>